# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Srulik's lab API
  version: 0.1.0
  description: >
    HTTP API for the Srulik's lab backend. This spec is the single source of truth for
    the generated API reference.
servers:
  - url: /
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      additionalProperties: true
      properties:
        error: { type: string }
        code: { type: string }
...
    SocialMediaFetchTopicRequest:
      type: object
      required: [topic]
      properties:
        topic:
          type: string
          minLength: 2
          description: Free-text topic or concept to search across platforms
        platforms:
          type: array
          items: { type: string, enum: [x, telegram, facebook] }
          description: Subset of platforms; at least two required when execute is true
        execute:
          type: boolean
          default: false
          description: When false, returns dry-run estimate only
        maxCostUsd:
          type: number
          description: Optional cost cap for live fetch
        maxPerQuery:
          type: integer
          description: Optional per-query result cap
        lang:
          type: string
          description: Optional UI language for post translation (e.g. en, he, ru)
    SocialMediaTopicFetchHistoryResponse:
      type: object
      required: [searches]
      properties:
        searches:
          type: array
          items: { type: object, additionalProperties: true }
...
paths:
...
  /api/education-sessions:
    get:
      tags: [Education]
      summary: Get education sessions dashboard.
      parameters:
        - in: query
          name: refresh
          schema: { type: string, enum: ["0", "1"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "502":
          description: Upstream failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/naftali:
    get:
      tags: [Naftali]
      summary: Get Naftali dashboard.
      parameters:
        - in: query
          name: refresh
          schema: { type: string, enum: ["0", "1"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "502":
          description: Upstream failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/social-media:
    get:
      tags: [SocialMedia]
      summary: Social OSINT dashboard — available bundle dates and counts.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "502":
          description: Upstream failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Social media service not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/social-media/platforms:
    get:
      tags: [SocialMedia]
      summary: Selectable platforms for topic fetch (X, Telegram, Facebook).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "503":
          description: Social media service not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/social-media/daily:
    get:
      tags: [SocialMedia]
      summary: Daily OSINT feed — findings grouped by emergency category.
      parameters:
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
          description: Bundle date (YYYY-MM-DD)
        - in: query
          name: category
          schema: { type: string }
          description: Optional emergency category id filter
        - in: query
          name: lang
          schema: { type: string }
          description: Optional UI language for post translation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          description: Missing or invalid date
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Daily feed not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: Upstream failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Social media service not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/social-media/fetch-topic:
    post:
      tags: [SocialMedia]
      summary: On-demand topic search across social platforms.
      description: >
        Dry-run by default (`execute: false`). When `execute` is true, at least two
        platforms must be selected. Results may be persisted for history retrieval.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SocialMediaFetchTopicRequest" }
      responses:
        "200":
          description: Topic fetch result (dry-run estimate or live posts)
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          description: Invalid topic or platform selection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: Upstream failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Social media service not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/social-media/topic-fetches:
    get:
      tags: [SocialMedia]
      summary: List recent on-demand topic searches.
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SocialMediaTopicFetchHistoryResponse" }
        "502":
          description: Upstream failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Social media service not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/social-media/topic-fetches/{id}:
    get:
      tags: [SocialMedia]
      summary: Load a saved topic fetch by id.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: lang
          schema: { type: string }
          description: Optional UI language for post translation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "404":
          description: Topic fetch not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: Upstream failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Social media service not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/social-media/report:
    get:
      tags: [SocialMedia]
      summary: OSINT bundle and markdown report for a date.
      parameters:
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          description: Missing or invalid date
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Report not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: Upstream failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Social media service not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...